requirement.txt
opencv-python
opencv-python-headless
pydub
minio
kafka-python
transformers
torch
whisper
fer
deepface
TikTokApi
numpy<2.0.0
imageio
ffmpeg-python
python-dotenv
requests
pytube
moviepy
fastapi
torchaudio
pyspark==3.3.0
mtcnn
uvicorn
yt-dlp
ffmpeg
tensorflow==2.10.1
python-multipart
jinja2==3.1.2
scipy
speechbrain
emoji
langdetect
sentencepiece==0.1.97
protobuf==3.19.4
huggingface_hub[hf_xet]
openai
python-dotenv